Split admin entry point between dev and prod
