+++ packages/strapi-admin/admin/src/renderApp.js
@@ -1,0 +1,26 @@
+/**
+ * Common configuration for the app in both dev an prod mode
+ */
+
+import { Provider } from 'react-redux';
+import React from 'react';
+import ReactDOM from 'react-dom';
+import { ConnectedRouter } from 'react-router-redux';
+import LanguageProvider from 'containers/LanguageProvider';
+import App from 'containers/App';
+import { history, store } from './createStore';
+
+const render = (translatedMessages) => {
+  ReactDOM.render(
+    <Provider store={store}>
+      <LanguageProvider messages={translatedMessages}>
+        <ConnectedRouter history={history}>
+          <App />
+        </ConnectedRouter>
+      </LanguageProvider>
+    </Provider>,
+    document.getElementById('app')
+  );
+};
+
+export default render;
