+++ packages/strapi-admin/package.json
