+++ packages/strapi-plugin-content-manager/package.json
