+++ packages/strapi-plugin-content-type-builder/package.json
