+++ packages/strapi-plugin-email/package.json
