+++ packages/strapi-plugin-graphql/package.json
