+++ packages/strapi-plugin-settings-manager/package.json
