+++ packages/strapi-plugin-upload/package.json
