+++ packages/strapi-plugin-users-permissions/package.json
