+++ packages/strapi/bin/strapi.js
